# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "api.ErrorResponse": {
                "properties": {
                    "error": {
                        "description": "Error is a short, human-readable description of what went\nwrong. The server never returns raw stack traces or DB\nerror strings here; the goal is to be informative without\nleaking internals.",
                        "example": "Invalid request payload",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "model.MediaFile": {
                "properties": {
                    "createdAt": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "size": {
                        "type": "integer"
                    },
                    "type": {
                        "description": "image/video etc.",
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "settings.ThemeUploadResponse": {
                "properties": {
                    "message": {
                        "example": "Theme uploaded successfully",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "upload.FilesListResponse": {
                "properties": {
                    "files": {
                        "items": {
                            "$ref": "#/components/schemas/model.MediaFile"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
            "upload.MessageResponse": {
                "properties": {
                    "message": {
                        "example": "File deleted",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "upload.MultiUploadResponse": {
                "properties": {
                    "files": {
                        "items": {
                            "$ref": "#/components/schemas/model.MediaFile"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "message": {
                        "example": "File upload completed",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "upload.UploadResponse": {
                "properties": {
                    "file": {
                        "$ref": "#/components/schemas/model.MediaFile"
                    },
                    "message": {
                        "example": "File uploaded successfully",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "JWT bearer token issued by /api/auth/login",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "contact": {
            "name": "GitHub Issues",
            "url": "https://github.com/vexgo-org/vexgo/issues"
        },
        "description": "Self-hosted blog CMS HTTP API.",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "termsOfService": "https://github.com/vexgo-org/vexgo",
        "title": "VexGo API",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/config/theme/upload": {
            "post": {
                "description": "Accepts a multipart/form-data body with a single\n'theme' part. The zip must contain a vexgo-theme.json\nmetadata file (either at the root or inside a single\nsubdirectory). Existing themes with the same id are\noverwritten.",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "title": "theme",
                                "type": "file"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    },
                    "description": "theme zip archive",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/settings.ThemeUploadResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "invalid zip / missing metadata / zip-slip entry"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Upload a theme zip",
                "tags": [
                    "config"
                ]
            }
        },
...
        "/upload": {
            "post": {
                "description": "Accepts a multipart/form-data body with a single 'file'\npart. The stored filename is a freshly generated UUID\nplus the sanitized extension from the original filename;\nany untrusted characters are stripped before persistence.\nRequires authentication.",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "title": "file",
                                "type": "file"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    },
                    "description": "file to upload",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/upload.UploadResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "missing or malformed form"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "failed to persist file"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Upload a single file",
                "tags": [
                    "uploads"
                ]
            }
        },
        "/upload/multiple": {
            "post": {
                "description": "Accepts a multipart/form-data body with one or more\n'files' parts. Per-file failures are silently dropped;\nthe response only includes the rows that persisted.\nRequires authentication.",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "title": "files",
                                "type": "file"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    },
                    "description": "files to upload (repeatable)",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/upload.MultiUploadResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "missing or malformed form"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Upload multiple files",
                "tags": [
                    "uploads"
                ]
            }
        },
        "/upload/my": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/upload.FilesListResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List the authenticated user's uploads",
                "tags": [
                    "uploads"
                ]
            }
        },
        "/upload/{id}": {
            "delete": {
                "description": "Only the original uploader or an admin may delete a file.\nThe id is the MediaFile.ID from the upload response.",
                "parameters": [
                    {
                        "description": "file id (UUID filename without extension)",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/upload.MessageResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "not authorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "file not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Delete an uploaded file",
                "tags": [
                    "uploads"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
...
}